backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
langchain>=0.3.0
langchain-community>=0.3.0
dashscope>=1.20.0
sse-starlette>=1.8.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
python-multipart>=0.0.6
aiofiles>=23.0.0
